+++ lib/twenv/commands/write_tweet.rb
@@ -30,15 +30,19 @@ class TWEnv::WriteTweet < TWEnv::Command
 
   # Delay using a 24-hour clock
   write-tweet --delay 19:00
+
+  # Delay a tweet until Christmas Eve morning, 2022
+  write-tweet --delay 09:00AM --delay-date 24/12/2022
   BANNER
 
   ONE_DAY = 3600 * 24
 
   def options(slop)
-    slop.on :d, :delay=         , 'Delay sending a tweet', as: :string, default: '0'
-    slop.on :f, :files=         , 'List of files to post with the tweet', as: :array, default: []
-    slop.on :r, 'in-reply-to='  , 'Write a reply to the given tweet', as: :boolean, default: nil
-    slop.on :s, 'show-schedule' , 'Show at what time delayed tweet(s) are scheduled to be published', as: :boolean, default: nil
+    slop.on :d,  :delay=         , 'Delay sending a tweet', as: :string, default: '0'
+    slop.on :f,  :files=         , 'List of files to post with the tweet', as: :array, default: []
+    slop.on :r,  'in-reply-to='  , 'Write a reply to the given tweet', as: :boolean, default: nil
+    slop.on :s,  'show-schedule' , 'Show at what time delayed tweet(s) are scheduled to be published', as: :boolean, default: nil
+    slop.on      'delay-date='   , 'The date to which the --delay option should be relative to. Defaults to today', as: :string, default: nil
   end
 
   def process
@@ -129,12 +133,13 @@ class TWEnv::WriteTweet < TWEnv::Command
   end
 
   def make_time(hour, minute, median)
+    delay_date = opts['delay-date'] || Date.today.strftime('%d/%m/%Y')
     if median
       # 12 hour clock
-      time_obj = Time.strptime("#{hour}:#{minute} #{median.downcase}", "%I:%M %P")
+      time_obj = Time.strptime("#{hour}:#{minute} #{median.downcase} #{delay_date}", "%I:%M %P %d/%m/%Y")
     else
       # 24 hour clock
-      time_obj = Time.strptime("#{hour}:#{minute}", "%H:%M")
+      time_obj = Time.strptime("#{hour}:#{minute} #{delay_date}", "%H:%M %d/%m/%Y")
     end
     time_obj < Time.now ? time_obj + ONE_DAY : time_obj
   rescue ArgumentError