+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## HEAD
 
+* Perform a client-side verification on tweet size before attempting
+  to send the tweet to Twitter. Tweets who exceed the maximum length
+  can be edited until their size fits within the contraints.
+
 * Add `TWEnv::WriteTweet::DelayedTweet` struct.
 
 * Add `TWENV::Line#info`.