+++ CHANGELOG.md
@@ -2,13 +2,16 @@
 
 ## HEAD
 
+* Show more information about scheduled tweets when using the 
+  `--show-schedule` option of the `write-tweet` command.
+
 * Add `--before-date=` option to the `delete-my-tweets` command.
 
 * Perform a client-side verification on tweet size before attempting
   to send the tweet to Twitter. Tweets who exceed the maximum length
   can be edited until their size fits within Twitter's size constraint.
 
-* Add `TWEnv::WriteTweet::DelayedTweet` struct.
+* Add `TWEnv::WriteTweet::ScheduledTweet` struct.
 
 * Add `TWENV::Line#info`.