+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
-## HEAD
+## Next release
+
+* Remove the `read-links` command. 
 
 * In the `delete-my-tweets` command, parse the `--before-date` option as a date 
   in iso8601 format.